requirements.txt
Flask==3.0.0
Werkzeug==3.0.1
pydantic==2.5.0
redis==5.0.1
Flask-Limiter==3.3.1
Flask-CORS==4.0.0
Flask-WTF==1.2.1
Flask-SocketIO==5.3.4
python-dotenv==1.0.0
python-engineio==4.7.0
python-socketio==5.9.0
gunicorn==22.0.0
Werkzeug[watchdog]==3.0.1
gevent==23.9.1
gevent-websocket==0.10.1